Don't use /usr/bin/python2, solves #24.
Replace it with /usr/bin/env that is more virtualenv-friendly. Basically
I had ~/virtualenv directory that I activated using bin/activate and
this setup didn't solve that. Note that requirements.txt had nothing to
do with it. I found that by importing the module manually in Python
interpreter and seeing that work.